Update LuaStandaloneInterpreter.php
Replaced
if ( php_uname( 's' ) == 'Windows NT' ) {
			// Like the passthru() in older versions of PHP,
			// PHP's invokation of cmd.exe in proc_open() is broken:
			// http://news.php.net/php.internals/21796
			// Unlike passthru(), it is not fixed in any PHP version,
			// so we use the fix similar to one in wfShellExec()
			$cmd = '"' . $cmd . '"';
		}


The above code can trigger an error In the Windows Environment This error being
"Lua error: Internal error: The interpreter exited with status 1."
 New Code:

if (php_uname('s') == 'Windows NT') {
    $cmd = escapeshellcmd($cmd);
}


escapeshellcmd($cmd):
The escapeshellcmd() function is used to escape any characters in a string that can be misinterpreted as shell metacharacters. It makes the string safe for passing to a shell command, which helps prevent command injection attacks. On Windows, this function also replaces % with space. Ideally eliminating double quotes entirelly

$cmd = escapeshellcmd($cmd);:
This line takes the original $cmd string, escapes it using escapeshellcmd(), and then assigns the escaped string back to $cmd

The code should escape any potentially dangerous characters in the $cmd variable to make it safe for execution as a shell command.